requirements.txt
psycopg2-binary==2.9.10
python-telegram-bot==21.7
